feat: um

+++ scripts/remove-all-symlinks.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/bin/sh
 
 # .config
 unlink ~/.config/nvim
@@ -6,6 +6,14 @@ unlink ~/.config/git
 
 # .home
 unlink ~/.tmux.conf
-unlink ~/.wezterm.lua
 unlink ~/.zshrc
 unlink ~/.p10k.zsh
+
+if [ "$WSL_DISTRO_NAME" = "Ubuntu" ];
+then
+  echo "setting up symlink for WSL"
+  # unlink ~/.wezterm.lua
+else
+  echo "setting up symlink for native linux"
+  unlink ~/.wezterm.lua
+fi
